move course enrolments bus to app level
