+++ app/src/main/java/org/stepic/droid/di/course_general/CourseEnrollmentBusModule.kt
@@ -2,6 +2,10 @@ package org.stepic.droid.di.course_general
 
 import dagger.Binds
 import dagger.Module
+import dagger.Provides
+import io.reactivex.Observable
+import io.reactivex.Scheduler
+import io.reactivex.subjects.PublishSubject
 import org.stepic.droid.base.Client
 import org.stepic.droid.base.ClientImpl
 import org.stepic.droid.base.ListenerContainer
@@ -13,33 +17,55 @@ import org.stepic.droid.core.joining.JoiningPosterImpl
 import org.stepic.droid.core.joining.contract.JoiningListener
 import org.stepic.droid.core.joining.contract.JoiningPoster
 import org.stepic.droid.di.AppSingleton
+import org.stepic.droid.di.qualifiers.BackgroundScheduler
+import org.stepik.android.model.Course
+import org.stepik.android.view.injection.course.EnrollmentCourseUpdates
 
 @Module
-interface CourseEnrollmentBusModule {
+internal abstract class CourseEnrollmentBusModule {
 
     @Binds
     @AppSingleton
-    fun bindsClient(clientImpl: ClientImpl<DroppingListener>): Client<DroppingListener>
+    internal abstract fun bindsClient(clientImpl: ClientImpl<DroppingListener>): Client<DroppingListener>
 
     @Binds
     @AppSingleton
-    fun bindContainer(listenerContainer: ListenerContainerImpl<DroppingListener>): ListenerContainer<DroppingListener>
+    internal abstract fun bindContainer(listenerContainer: ListenerContainerImpl<DroppingListener>): ListenerContainer<DroppingListener>
 
     @Binds
     @AppSingleton
-    fun bindsPoster(posterImpl: DroppingPosterImpl): DroppingPoster
+    internal abstract fun bindsPoster(posterImpl: DroppingPosterImpl): DroppingPoster
 
 
     @Binds
     @AppSingleton
-    fun bindsJoinClient(clientImpl: ClientImpl<JoiningListener>): Client<JoiningListener>
+    internal abstract fun bindsJoinClient(clientImpl: ClientImpl<JoiningListener>): Client<JoiningListener>
 
     @Binds
     @AppSingleton
-    fun bindJoinContainer(listenerContainer: ListenerContainerImpl<JoiningListener>): ListenerContainer<JoiningListener>
+    internal abstract fun bindJoinContainer(listenerContainer: ListenerContainerImpl<JoiningListener>): ListenerContainer<JoiningListener>
 
     @Binds
     @AppSingleton
-    fun bindsJoinPoster(posterImpl: JoiningPosterImpl): JoiningPoster
+    internal abstract fun bindsJoinPoster(posterImpl: JoiningPosterImpl): JoiningPoster
 
+    @Module
+    companion object {
+        @Provides
+        @JvmStatic
+        @AppSingleton
+        @EnrollmentCourseUpdates
+        internal fun provideCourseEnrollmentSubject(): PublishSubject<Course> =
+            PublishSubject.create()
+
+        @Provides
+        @JvmStatic
+        @AppSingleton
+        @EnrollmentCourseUpdates
+        internal fun bindEnrollmentsUpdatesObservables(
+            @EnrollmentCourseUpdates publisher: PublishSubject<Course>,
+            @BackgroundScheduler scheduler: Scheduler
+        ): Observable<Course> =
+            publisher.observeOn(scheduler)
+    }
 }
