+++ app/src/main/java/org/stepik/android/view/injection/course/CourseModule.kt
@@ -8,7 +8,6 @@ import dagger.multibindings.IntoMap
 import io.reactivex.Observable
 import io.reactivex.Scheduler
 import io.reactivex.subjects.BehaviorSubject
-import io.reactivex.subjects.PublishSubject
 import org.stepic.droid.di.qualifiers.BackgroundScheduler
 import org.stepik.android.model.Course
 import org.stepik.android.presentation.base.injection.ViewModelKey
@@ -55,22 +54,5 @@ abstract class CourseModule {
         @CourseScope
         internal fun provideCourseObservableSource(courseSubject: BehaviorSubject<Course>, @BackgroundScheduler scheduler: Scheduler): Observable<Course> =
             courseSubject.observeOn(scheduler)
-
-        @Provides
-        @JvmStatic
-        @CourseScope
-        @EnrollmentCourseUpdates
-        internal fun provideCourseEnrollmentSubject(): PublishSubject<Course> =
-            PublishSubject.create()
-
-        @Provides
-        @JvmStatic
-        @CourseScope
-        @EnrollmentCourseUpdates
-        internal fun bindEnrollmentsUpdatesObservables(
-            @EnrollmentCourseUpdates publisher: PublishSubject<Course>,
-            @BackgroundScheduler scheduler: Scheduler
-        ): Observable<Course> =
-            publisher.observeOn(scheduler)
     }
 }
